Split channel_drawer into multiple components
